Removing leading and trailing quotes from result. Fix for issue #20.
False positives are possible (the quotes are actually in the source
string and not added by reading stdout), but I don't know how it could
be done more reliably and the error rate should be rather low.
A workaround for false positives could be to add extra quotes at the
beginning and at the end of the source string, e.g. ""This string will
be quoted with a single quote character"" or using single quotes instead
of double quotes.